Interactions with brick and cat working
Interactions and collisions working and with sound and visual for bricks interaction.

Added script so that the door breaks after a few hits with bricks.

Still need to correct the cats interactions a bit (making him non kinematic but constrained in rotation so that he wont fall.

Also need to make crawlers scarier

And lastly i need to add UI with poke interactions. Like for starters some instructions